% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep_jhu.R
\name{jhu_sitrep_import}
\alias{jhu_sitrep_import}
\alias{jhu_sitrep_cleandb}
\title{jhu sitrep data management and visualization}
\usage{
jhu_sitrep_import(source)

jhu_sitrep_cleandb(data)
}
\arguments{
\item{source}{string of source to use: "confirmed","deaths" or "recovered"}

\item{data}{input of raw jhu dataset}
}
\value{
import and cleaned jhu dataset
}
\description{
clean and plot jhu sitrep
}
\section{Functions}{
\itemize{
\item \code{jhu_sitrep_import}: import data from github

\item \code{jhu_sitrep_cleandb}: clean jhu dataset
}}

\examples{

library(covid19viz)
library(tidyverse)

jhu_sitrep <- jhu_sitrep_import(source = "confirmed")

jhu_sitrep \%>\%
  jhu_sitrep_cleandb() \%>\%
  filter(country_region=="Peru") \%>\%
  arrange(desc(dates))

}
