% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_intervention.R
\name{read_google_global}
\alias{read_google_global}
\alias{read_google_region_list}
\alias{read_google_region_country}
\alias{read_acaps_governments}
\alias{read_unesco_education}
\title{import mobility and intervention global data}
\usage{
read_google_global()

read_google_region_list()

read_google_region_country(country_iso = "PE")

read_acaps_governments()

read_unesco_education()
}
\arguments{
\item{country_iso}{two letter iso code for each country available in read_google_region_list()}
}
\value{
import external data sources on covid19 human behaviour or interventions
}
\description{
import mobility google reports, acasp and unesco data on interventions
}
\section{Functions}{
\itemize{
\item \code{read_google_global}: import google mobility reports

\item \code{read_google_region_list}: print list of files inside regional zip file

\item \code{read_google_region_country}: import regional mobility data for one country

\item \code{read_acaps_governments}: import acaps dataset

\item \code{read_unesco_education}: import unesco dataset
}}

\examples{

\dontrun{

library(covid19viz)
library(tidyverse)

# google mobility reports

# global
read_google_global()

# regional
# first: select country ISO
read_google_region_list()

# second: read specific country data
peru <- read_google_region_country(country_iso = "PE")
peru \%>\% count(sub_region_1)

# acaps data
read_acaps_governments()

# unesco education data
read_unesco_education()

}

}
