% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep_who.R
\name{who_sitrep_import}
\alias{who_sitrep_import}
\alias{who_sitrep_cleandb}
\alias{who_sitrep_ggline}
\alias{who_sitrep_ggbar}
\alias{who_sitrep_country_report}
\title{who sitrep data management and visualization}
\usage{
who_sitrep_import(update)

who_sitrep_cleandb(data)

who_sitrep_ggline(data, country = "Peru", y_cum_value, color,
  n_breaks = 5)

who_sitrep_ggbar(data, country = "Peru", y_inc_value, fill,
  n_breaks = 5)

who_sitrep_country_report(update, country)
}
\arguments{
\item{update}{string of date to update}

\item{data}{input of raw who dataset}

\item{country}{string of country name}

\item{y_cum_value}{cumulative variable}

\item{color}{colo of attribute}

\item{n_breaks}{y axis breaks}

\item{y_inc_value}{incidence variable}

\item{fill}{fill of attribute}
}
\value{
report per country using digitalized who sitreps
}
\description{
clean and plot who sitrep
}
\section{Functions}{
\itemize{
\item \code{who_sitrep_import}: import data from github

\item \code{who_sitrep_cleandb}: clean who dataset

\item \code{who_sitrep_ggline}: plot a ggplot geom_line

\item \code{who_sitrep_ggbar}: plot a ggplot geom_col

\item \code{who_sitrep_country_report}: create a full unified report
}}

\examples{

library(covid19viz)
library(tidyverse)

who_sitrep <- who_sitrep_import(update = "2020-03-10")

who_sitrep \%>\%
    who_sitrep_cleandb() \%>\%
    filter(country=="Peru")

who_sitrep \%>\%
    who_sitrep_cleandb() \%>\%
    who_sitrep_ggline(country = "Peru",
                      y_cum_value = n_cum_conf,
                      color = class, n_breaks = 10)

}
