% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep_who.R
\name{who_sitrep_import}
\alias{who_sitrep_import}
\alias{who_sitrep_cleandb}
\alias{who_sitrep_ggline}
\alias{who_sitrep_ggbar}
\alias{who_sitrep_country_report}
\title{who sitrep data management}
\usage{
who_sitrep_import(data_input)

who_sitrep_cleandb(data)

who_sitrep_ggline(data, country = "Peru", y_cum_value, color,
  n_breaks = 5)

who_sitrep_ggbar(data, country = "Peru", y_inc_value, fill,
  n_breaks = 5)

who_sitrep_country_report(data_input, country)
}
\arguments{
\item{data_input}{string of github repo}

\item{data}{input of raw dataset}

\item{country}{string of country name}

\item{y_cum_value}{cumulative variable}

\item{color}{colo of attribute}

\item{n_breaks}{y axis breaks}

\item{y_inc_value}{incidence variable}

\item{fill}{fill of attribute}
}
\value{
report per country
}
\description{
clean and plot who sitrep
}
\section{Functions}{
\itemize{
\item \code{who_sitrep_import}: import data from github

\item \code{who_sitrep_cleandb}: clean dataset

\item \code{who_sitrep_ggline}: plot a ggplot geom_line

\item \code{who_sitrep_ggbar}: plot a ggplot geom_col

\item \code{who_sitrep_country_report}: remove replicates
}}

\examples{

who_sitrep_country_report(data_input = "https://raw.github.com/fkrauer/COVID-19/master/data/WHO_COVID19_ALL_ADM0_2020-03-10.csv",
                          country = "Brazil")

}
